Refactor TimeReport and related modules for better readability and maintainability
- Renamed `summary` to `report` in `application.rb` for clarity.
- Removed the `Formatter` module and integrated its functionality into `Table` and `TimeBlockChart`.
- Refactored `Table` module to improve method clarity and reduce redundancy.
- Added new methods in `Table` for formatting specific parts of the table row.
- Enhanced `TimeBlockChart` with new methods for formatting and printing date information.
- Integrated `Table` and `TimeBlockChart` into `TimeReport` for a more modular structure.
- Removed redundant methods from `TimeReportHelper` and ensured all necessary methods are included in the appropriate modules.
